pyproject.toml
[project]
name = "prismo"
version = "0.1.0"
description = "A microscopy image acquisition toolkit."
authors = [
    {name = "Karl Krauth", email = "karl.krauth@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12, <3.14"

dependencies = [
    "multiprocess>=0.70.0",
    "napari>=0.5.0",
    "pybind11>=2.13.5",
    "pymmcore",
    "pymodbus>=3.9.0",
    "pyserial>=3.5",
    "xarray[io]>=2025.01.0",
    "zarr>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
    "ty>=0.0.1a24",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true

[build-system]
requires = ["uv_build>=0.9.5"]
build-backend = "uv_build"

